+++ next.config.js
@@ -1,0 +1,9 @@
+/** @type {import('next').NextConfig} */
+// const withBundleAnalyzer = require('@next/bundle-analyzer')({
+//     enabled: process.env.ANALYZE === 'true',
+// });
+
+const nextConfig = {};
+// module.exports = withBundleAnalyzer(nextConfig);
+
+// export default nextConfig;
