+++ package.json
@@ -10,8 +10,10 @@
   },
   "dependencies": {
     "axios": "^1.6.7",
+    "dotenv": "^16.4.5",
+    "express": "^4.18.3",
     "history": "^5.3.0",
-    "next": "14.1.0",
+    "next": "^14.1.0",
     "next-auth": "^4.24.6",
     "react": "^18",
     "react-dom": "^18",
