+++ src/app/components/Login/index.tsx
@@ -59,16 +59,18 @@ const LoginComponent = () => {
     }
   };
 
-  if (session) {
-    router.push("/dashboard");
-  }
+  useEffect(() => {
+    if (session !== "") {
+      router.push("/dashboard");
+    }
+  }, [session, router]);
 
   if (!user) return null;
   return (
     <>
       <div className="container mx-auto px-4 mt-4">
         <div className="flex justify-center h-full">
-          <div className="w-full lg:w-4/12 px-4 ">
+          <div className="w-full lg:w-4/12 px-4">
             <div className="relative bg-blue-50 flex flex-col min-w-0 break-words w-full mb-6 shadow-lg rounded-lg bg-blueGray-200 border-0">
               <div className="rounded-t mb-0 px-6 py-6">
                 <div className="text-center mb-3">
@@ -87,7 +89,7 @@ const LoginComponent = () => {
                   <div className="relative w-full">
                     <label
                       className="block uppercase text-blueGray-600 text-xs font-bold mb-2"
-                      htmlFor="grid-password"
+                      htmlFor="EmailAddress"
                     >
                       Email*
                     </label>
@@ -98,6 +100,8 @@ const LoginComponent = () => {
                       className="border-0 px-3 py-3 placeholder-blueGray-300 text-blueGray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
                       placeholder="Email"
                       name="email"
+                      id="EmailAddress"
+                      autoComplete="off"
                     />
                   </div>
                   <div className="">
