+++ src/app/config.ts
@@ -1,0 +1,1 @@
+export const ServerPort = process.env.NEXT_AUTH_URL;
