--- src/components/footer/Footer.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Footer = () => {
-  return <div>Footer</div>;
-};
-
-export default Footer;
